+++ libs/error-stack/src/result.rs
@@ -1,54 +1,11 @@
-#![expect(deprecated, reason = "We use `Context` to maintain compatibility")]
+use core::error::Error;
 
-use crate::{Attachment, Context, IntoReport, OpaqueAttachment, Report};
+use crate::{Attachment, IntoReport, OpaqueAttachment, Report};
 
-#[expect(rustdoc::invalid_html_tags, reason = "False positive")]
-/// <code>[Result](core::result::Result)<T, [Report<C>]></code>
-///
-/// A reasonable return type to use throughout an application.
-///
-/// The `Result` type can be used with one or two parameters, where the first parameter represents
-/// the [`Ok`] arm and the second parameter `Error` is used as in [`Report<C>`].
-///
-/// # Examples
-///
-/// `Result` can also be used in `fn main()`:
-///
-/// ```rust
-/// # fn has_permission(_: (), _: ()) -> bool { true }
-/// # fn get_user() -> Result<(), AccessError> { Ok(()) }
-/// # fn get_resource() -> Result<(), Report<AccessError>> { Ok(()) }
-/// # #[derive(Debug)] enum AccessError { PermissionDenied((), ()) }
-/// # impl core::fmt::Display for AccessError {
-/// #    fn fmt(&self, _: &mut std::fmt::Formatter<'_>) -> std::fmt::Result { Ok(()) }
-/// # }
-/// # impl core::error::Error for AccessError {}
-/// use error_stack::{ensure, Report};
-///
-/// fn main() -> Result<(), Report<AccessError>> {
-///     let user = get_user()?;
-///     let resource = get_resource()?;
-///
-///     ensure!(
-///         has_permission(user, resource),
-///         AccessError::PermissionDenied(user, resource)
-///     );
-///
-///     # const _: &str = stringify! {
-///     ...
-///     # }; Ok(())
-/// }
-/// ```
-#[deprecated(
-    note = "Use `core::result::Result<T, Report<C>>` instead",
-    since = "0.6.0"
-)]
-pub type Result<T, C> = core::result::Result<T, Report<C>>;
-
-/// Extension trait for [`Result`][core::result::Result] to provide context information on
+/// Extension trait for [`Result`] to provide context information on
 /// [`Report`]s.
 pub trait ResultExt {
-    /// The [`Context`] type of the [`Result`].
+    /// The context type of the [`Result`] which becomes the `C` in `Report<C>`.
     type Context: ?Sized;
 
     /// Type of the [`Ok`] value in the [`Result`]
@@ -58,7 +15,7 @@ pub trait ResultExt {
     ///
     /// Applies [`Report::attach`] on the [`Err`] variant, refer to it for more
     /// information.
-    fn attach<A>(self, attachment: A) -> core::result::Result<Self::Ok, Report<Self::Context>>
+    fn attach<A>(self, attachment: A) -> Result<Self::Ok, Report<Self::Context>>
     where
         A: Attachment;
 
@@ -66,90 +23,43 @@ pub trait ResultExt {
     ///
     /// Applies [`Report::attach`] on the [`Err`] variant, refer to it for more
     /// information.
-    fn attach_with<A, F>(
-        self,
-        attachment: F,
-    ) -> core::result::Result<Self::Ok, Report<Self::Context>>
+    fn attach_with<A, F>(self, attachment: F) -> Result<Self::Ok, Report<Self::Context>>
     where
         A: Attachment,
-        F: FnOnce() -> A;
-
-    #[deprecated(
-        note = "Use `attach_opaque_with` instead. `attach_lazy` was renamed to \
-                `attach_opaque_with` and `attach_printable_lazy` was renamed to `attach_with`",
-        since = "0.6.0"
-    )]
-    fn attach_lazy<A, F>(
-        self,
-        attachment: F,
-    ) -> core::result::Result<Self::Ok, Report<Self::Context>>
-    where
-        A: OpaqueAttachment,
         F: FnOnce() -> A;
 
     /// Adds a new attachment to the [`Report`] inside the [`Result`].
     ///
     /// Applies [`Report::attach_opaque`] on the [`Err`] variant, refer to it for more information.
-    fn attach_opaque<A>(
-        self,
-        attachment: A,
-    ) -> core::result::Result<Self::Ok, Report<Self::Context>>
+    fn attach_opaque<A>(self, attachment: A) -> Result<Self::Ok, Report<Self::Context>>
     where
         A: OpaqueAttachment;
 
     /// Lazily adds a new attachment to the [`Report`] inside the [`Result`].
     ///
     /// Applies [`Report::attach_opaque`] on the [`Err`] variant, refer to it for more information.
-    fn attach_opaque_with<A, F>(
-        self,
-        attachment: F,
-    ) -> core::result::Result<Self::Ok, Report<Self::Context>>
+    fn attach_opaque_with<A, F>(self, attachment: F) -> Result<Self::Ok, Report<Self::Context>>
     where
         A: OpaqueAttachment,
-        F: FnOnce() -> A;
-
-    #[deprecated(
-        note = "Use `attach` instead. `attach` was renamed to `attach_opaque` and \
-                `attach_printable` was renamed to `attach`",
-        since = "0.6.0"
-    )]
-    fn attach_printable<A>(
-        self,
-        attachment: A,
-    ) -> core::result::Result<Self::Ok, Report<Self::Context>>
-    where
-        A: Attachment;
-
-    #[deprecated(
-        note = "Use `attach_with` instead. `attach_lazy` was renamed to `attach_opaque_with` and \
-                `attach_printable_lazy` was renamed to `attach_with`",
-        since = "0.6.0"
-    )]
-    fn attach_printable_lazy<A, F>(
-        self,
-        attachment: F,
-    ) -> core::result::Result<Self::Ok, Report<Self::Context>>
-    where
-        A: Attachment,
         F: FnOnce() -> A;
 
     /// Changes the context of the [`Report`] inside the [`Result`].
     ///
     /// Applies [`Report::change_context`] on the [`Err`] variant, refer to it for more information.
-    fn change_context<C>(self, context: C) -> core::result::Result<Self::Ok, Report<C>>
+    fn change_context<C>(self, context: C) -> Result<Self::Ok, Report<C>>
     where
-        C: Context;
+        C: Error + Send + Sync + 'static;
 
     /// Lazily changes the context of the [`Report`] inside the [`Result`].
     ///
     /// Applies [`Report::change_context`] on the [`Err`] variant, refer to it for more information.
-    fn change_context_lazy<C, F>(self, context: F) -> core::result::Result<Self::Ok, Report<C>>
+    fn change_context_lazy<C, F>(self, context: F) -> Result<Self::Ok, Report<C>>
     where
-        C: Context,
+        C: Error + Send + Sync + 'static,
         F: FnOnce() -> C;
 }
 
-impl<T, E> ResultExt for core::result::Result<T, E>
+impl<T, E> ResultExt for Result<T, E>
 where
     E: IntoReport,
 {
@@ -157,7 +67,7 @@ where
     type Ok = T;
 
     #[track_caller]
-    fn attach<A>(self, attachment: A) -> core::result::Result<T, Report<E::Context>>
+    fn attach<A>(self, attachment: A) -> Result<T, Report<E::Context>>
     where
         A: Attachment,
     {
@@ -168,7 +78,7 @@ where
     }
 
     #[track_caller]
-    fn attach_with<A, F>(self, attachment: F) -> core::result::Result<T, Report<E::Context>>
+    fn attach_with<A, F>(self, attachment: F) -> Result<T, Report<E::Context>>
     where
         A: Attachment,
         F: FnOnce() -> A,
@@ -180,19 +90,7 @@ where
     }
 
     #[track_caller]
-    fn attach_lazy<A, F>(self, attachment: F) -> core::result::Result<T, Report<E::Context>>
-    where
-        A: OpaqueAttachment,
-        F: FnOnce() -> A,
-    {
-        match self {
-            Ok(value) => Ok(value),
-            Err(error) => Err(error.into_report().attach_opaque(attachment())),
-        }
-    }
-
-    #[track_caller]
-    fn attach_opaque<A>(self, attachment: A) -> core::result::Result<T, Report<E::Context>>
+    fn attach_opaque<A>(self, attachment: A) -> Result<T, Report<E::Context>>
     where
         A: OpaqueAttachment,
     {
@@ -203,7 +101,7 @@ where
     }
 
     #[track_caller]
-    fn attach_opaque_with<A, F>(self, attachment: F) -> core::result::Result<T, Report<E::Context>>
+    fn attach_opaque_with<A, F>(self, attachment: F) -> Result<T, Report<E::Context>>
     where
         A: OpaqueAttachment,
         F: FnOnce() -> A,
@@ -215,35 +113,9 @@ where
     }
 
     #[track_caller]
-    fn attach_printable<A>(self, attachment: A) -> core::result::Result<T, Report<E::Context>>
+    fn change_context<C>(self, context: C) -> Result<T, Report<C>>
     where
-        A: Attachment,
-    {
-        match self {
-            Ok(value) => Ok(value),
-            Err(error) => Err(error.into_report().attach(attachment)),
-        }
-    }
-
-    #[track_caller]
-    fn attach_printable_lazy<A, F>(
-        self,
-        attachment: F,
-    ) -> core::result::Result<T, Report<E::Context>>
-    where
-        A: Attachment,
-        F: FnOnce() -> A,
-    {
-        match self {
-            Ok(value) => Ok(value),
-            Err(error) => Err(error.into_report().attach(attachment())),
-        }
-    }
-
-    #[track_caller]
-    fn change_context<C>(self, context: C) -> core::result::Result<T, Report<C>>
-    where
-        C: Context,
+        C: Error + Send + Sync + 'static,
     {
         match self {
             Ok(value) => Ok(value),
@@ -252,9 +124,9 @@ where
     }
 
     #[track_caller]
-    fn change_context_lazy<C, F>(self, context: F) -> core::result::Result<T, Report<C>>
+    fn change_context_lazy<C, F>(self, context: F) -> Result<T, Report<C>>
     where
-        C: Context,
+        C: Error + Send + Sync + 'static,
         F: FnOnce() -> C,
     {
         match self {